split generic in base+generic
